Move all remaining APIs to `serialization.go`.
